+++ internal/home/mobileconfig.go
@@ -10,35 +10,60 @@ import (
 	"github.com/AdguardTeam/AdGuardHome/internal/dnsforward"
 	"github.com/AdguardTeam/golibs/errors"
 	"github.com/AdguardTeam/golibs/log"
+	"github.com/AdguardTeam/golibs/stringutil"
 	uuid "github.com/satori/go.uuid"
 	"howett.net/plist"
 )
 
+// dnsSettings is the DNSSetting.DNSSettings mobileconfig profile.
+//
+// See https://developer.apple.com/documentation/devicemanagement/dnssettings/dnssettings.
 type dnsSettings struct {
+	// DNSProtocol is the required protocol to be used.  The valid values
+	// are "HTTPS" and "TLS".
 	DNSProtocol string
-	ServerURL   string `plist:",omitempty"`
-	ServerName  string `plist:",omitempty"`
-	clientID    string
+
+	// ServerURL is the URI template of the DoH server.  It must be empty if
+	// DNSProtocol is not "HTTPS".
+	ServerURL string `plist:",omitempty"`
+
+	// ServerName is the hostname of the DoT server.  It must be empty if
+	// DNSProtocol is not "TLS".
+	ServerName string `plist:",omitempty"`
+
+	// ServerAddresses is a list of plain DNS server IP addresses used to
+	// resolve the hostname in ServerURL or ServerName.
+	ServerAddresses []string `plist:",omitempty"`
 }
 
+// payloadContent is a Device Management Profile payload.
+//
+// See https://developer.apple.com/documentation/devicemanagement/configuring_multiple_devices_using_profiles#3234127.
 type payloadContent struct {
-	Name               string
-	PayloadDescription string
-	PayloadDisplayName string
-	PayloadIdentifier  string
+	DNSSettings *dnsSettings
+
 	PayloadType        string
+	PayloadIdentifier  string
 	PayloadUUID        string
-	DNSSettings        dnsSettings
+	PayloadDisplayName string
+	PayloadDescription string
 	PayloadVersion     int
 }
 
+// dnsSettingsPayloadType is the payload type for a DNSSettings profile.
+const dnsSettingsPayloadType = "com.apple.dnsSettings.managed"
+
+// mobileConfig contains the TopLevel properties for configuring Device
+// Management Profiles.
+//
+// See https://developer.apple.com/documentation/devicemanagement/toplevel.
 type mobileConfig struct {
 	PayloadDescription       string
 	PayloadDisplayName       string
 	PayloadIdentifier        string
 	PayloadType              string
 	PayloadUUID              string
-	PayloadContent           []payloadContent
+	PayloadContent           []*payloadContent
 	PayloadVersion           int
 	PayloadRemovalDisallowed bool
 }
@@ -52,7 +77,7 @@ const (
 	dnsProtoTLS   = "TLS"
 )
 
-func getMobileConfig(d dnsSettings) ([]byte, error) {
+func encodeMobileConfig(d *dnsSettings, clientID string) ([]byte, error) {
 	var dspName string
 	switch proto := d.DNSProtocol; proto {
 	case dnsProtoHTTPS:
@@ -60,41 +85,41 @@ func getMobileConfig(d dnsSettings) ([]byte, error) {
 		u := &url.URL{
 			Scheme: schemeHTTPS,
 			Host:   d.ServerName,
-			Path:   path.Join("/dns-query", d.clientID),
+			Path:   path.Join("/dns-query", clientID),
 		}
 		d.ServerURL = u.String()
+
 		// Empty the ServerName field since it is only must be presented
 		// in DNS-over-TLS configuration.
-		//
-		// See https://developer.apple.com/documentation/devicemanagement/dnssettings/dnssettings.
 		d.ServerName = ""
 	case dnsProtoTLS:
 		dspName = fmt.Sprintf("%s DoT", d.ServerName)
-		if d.clientID != "" {
-			d.ServerName = d.clientID + "." + d.ServerName
+		if clientID != "" {
+			d.ServerName = clientID + "." + d.ServerName
 		}
 	default:
 		return nil, fmt.Errorf("bad dns protocol %q", proto)
 	}
 
-	data := mobileConfig{
-		PayloadContent: []payloadContent{{
-			Name:               dspName,
-			PayloadDescription: "Configures device to use AdGuard Home",
-			PayloadDisplayName: dspName,
-			PayloadIdentifier:  fmt.Sprintf("com.apple.dnsSettings.managed.%s", genUUIDv4()),
-			PayloadType:        "com.apple.dnsSettings.managed",
+	payloadID := fmt.Sprintf("%s.%s", dnsSettingsPayloadType, genUUIDv4())
+	data := &mobileConfig{
+		PayloadDescription: "Adds AdGuard Home to macOS Big Sur " +
+			"and iOS 14 or newer systems",
+		PayloadDisplayName: dspName,
+		PayloadIdentifier:  genUUIDv4(),
+		PayloadType:        "Configuration",
+		PayloadUUID:        genUUIDv4(),
+		PayloadContent: []*payloadContent{{
+			PayloadType:        dnsSettingsPayloadType,
+			PayloadIdentifier:  payloadID,
 			PayloadUUID:        genUUIDv4(),
+			PayloadDisplayName: dspName,
+			PayloadDescription: "Configures device to use AdGuard Home",
 			PayloadVersion:     1,
 			DNSSettings:        d,
 		}},
-		PayloadDescription:       "Adds AdGuard Home to Big Sur and iOS 14 or newer systems",
-		PayloadDisplayName:       dspName,
-		PayloadIdentifier:        genUUIDv4(),
-		PayloadRemovalDisallowed: false,
-		PayloadType:              "Configuration",
-		PayloadUUID:              genUUIDv4(),
 		PayloadVersion:           1,
+		PayloadRemovalDisallowed: false,
 	}
 
 	return plist.MarshalIndent(data, plist.XMLFormat, "\t")
@@ -133,13 +158,13 @@ func handleMobileConfig(w http.ResponseWriter, r *http.Request, dnsp string) {
 		}
 	}
 
-	d := dnsSettings{
-		DNSProtocol: dnsp,
-		ServerName:  host,
-		clientID:    clientID,
+	d := &dnsSettings{
+		DNSProtocol:     dnsp,
+		ServerName:      host,
+		ServerAddresses: cloneBootstrap(),
 	}
 
-	mobileconfig, err := getMobileConfig(d)
+	mobileconfig, err := encodeMobileConfig(d, clientID)
 	if err != nil {
 		respondJSONError(w, http.StatusInternalServerError, err.Error())
 
@@ -161,6 +186,14 @@ func handleMobileConfig(w http.ResponseWriter, r *http.Request, dnsp string) {
 	w.Header().Set("Content-Disposition", contDisp)
 
 	_, _ = w.Write(mobileconfig)
+}
+
+// cloneBootstrap returns a clone of the current bootstrap DNS servers.
+func cloneBootstrap() (bootstrap []string) {
+	config.RLock()
+	defer config.RUnlock()
+
+	return stringutil.CloneSlice(config.DNS.BootstrapDNS)
 }
 
 func handleMobileConfigDoH(w http.ResponseWriter, r *http.Request) {