+++ internal/home/mobileconfig_test.go
@@ -7,12 +7,39 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/AdguardTeam/AdGuardHome/internal/dnsforward"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"howett.net/plist"
 )
 
+// testBootstrapDNS are the bootstrap plain DNS server addresses for tests.
+var testBootstrapDNS = []string{
+	"94.140.14.14",
+	"94.140.15.15",
+}
+
+// setupBootstraps is a helper that sets up the bootstrap plain DNS server
+// configuration for tests and also tears it down in a cleanup function.
+func setupBootstraps(t testing.TB) {
+	t.Helper()
+
+	prevConfig := config
+	t.Cleanup(func() {
+		config = prevConfig
+	})
+	config = &configuration{
+		DNS: dnsConfig{
+			FilteringConfig: dnsforward.FilteringConfig{
+				BootstrapDNS: testBootstrapDNS,
+			},
+		},
+	}
+}
+
 func TestHandleMobileConfigDoH(t *testing.T) {
+	setupBootstraps(t)
+
 	t.Run("success", func(t *testing.T) {
 		r, err := http.NewRequest(http.MethodGet, "https://example.com:12345/apple/doh.mobileconfig?host=example.org", nil)
 		require.NoError(t, err)
@@ -25,11 +52,16 @@ func TestHandleMobileConfigDoH(t *testing.T) {
 		var mc mobileConfig
 		_, err = plist.Unmarshal(w.Body.Bytes(), &mc)
 		require.NoError(t, err)
-
 		require.Len(t, mc.PayloadContent, 1)
-		assert.Equal(t, "example.org DoH", mc.PayloadContent[0].Name)
+
 		assert.Equal(t, "example.org DoH", mc.PayloadContent[0].PayloadDisplayName)
-		assert.Equal(t, "https://example.org/dns-query", mc.PayloadContent[0].DNSSettings.ServerURL)
+
+		s := mc.PayloadContent[0].DNSSettings
+		require.NotNil(t, s)
+
+		assert.Equal(t, testBootstrapDNS, s.ServerAddresses)
+		assert.Empty(t, s.ServerName)
+		assert.Equal(t, "https://example.org/dns-query", s.ServerURL)
 	})
 
 	t.Run("error_no_host", func(t *testing.T) {
@@ -66,15 +98,22 @@ func TestHandleMobileConfigDoH(t *testing.T) {
 		var mc mobileConfig
 		_, err = plist.Unmarshal(w.Body.Bytes(), &mc)
 		require.NoError(t, err)
-
 		require.Len(t, mc.PayloadContent, 1)
-		assert.Equal(t, "example.org DoH", mc.PayloadContent[0].Name)
+
 		assert.Equal(t, "example.org DoH", mc.PayloadContent[0].PayloadDisplayName)
-		assert.Equal(t, "https://example.org/dns-query/cli42", mc.PayloadContent[0].DNSSettings.ServerURL)
+
+		s := mc.PayloadContent[0].DNSSettings
+		require.NotNil(t, s)
+
+		assert.Equal(t, testBootstrapDNS, s.ServerAddresses)
+		assert.Empty(t, s.ServerName)
+		assert.Equal(t, "https://example.org/dns-query/cli42", s.ServerURL)
 	})
 }
 
 func TestHandleMobileConfigDoT(t *testing.T) {
+	setupBootstraps(t)
+
 	t.Run("success", func(t *testing.T) {
 		r, err := http.NewRequest(http.MethodGet, "https://example.com:12345/apple/dot.mobileconfig?host=example.org", nil)
 		require.NoError(t, err)
@@ -87,11 +126,16 @@ func TestHandleMobileConfigDoT(t *testing.T) {
 		var mc mobileConfig
 		_, err = plist.Unmarshal(w.Body.Bytes(), &mc)
 		require.NoError(t, err)
-
 		require.Len(t, mc.PayloadContent, 1)
-		assert.Equal(t, "example.org DoT", mc.PayloadContent[0].Name)
+
 		assert.Equal(t, "example.org DoT", mc.PayloadContent[0].PayloadDisplayName)
-		assert.Equal(t, "example.org", mc.PayloadContent[0].DNSSettings.ServerName)
+
+		s := mc.PayloadContent[0].DNSSettings
+		require.NotNil(t, s)
+
+		assert.Equal(t, testBootstrapDNS, s.ServerAddresses)
+		assert.Equal(t, "example.org", s.ServerName)
+		assert.Empty(t, s.ServerURL)
 	})
 
 	t.Run("error_no_host", func(t *testing.T) {
@@ -129,10 +173,15 @@ func TestHandleMobileConfigDoT(t *testing.T) {
 		var mc mobileConfig
 		_, err = plist.Unmarshal(w.Body.Bytes(), &mc)
 		require.NoError(t, err)
-
 		require.Len(t, mc.PayloadContent, 1)
-		assert.Equal(t, "example.org DoT", mc.PayloadContent[0].Name)
+
 		assert.Equal(t, "example.org DoT", mc.PayloadContent[0].PayloadDisplayName)
-		assert.Equal(t, "cli42.example.org", mc.PayloadContent[0].DNSSettings.ServerName)
+
+		s := mc.PayloadContent[0].DNSSettings
+		require.NotNil(t, s)
+
+		assert.Equal(t, testBootstrapDNS, s.ServerAddresses)
+		assert.Equal(t, "cli42.example.org", s.ServerName)
+		assert.Empty(t, s.ServerURL)
 	})
 }